기병 돌격 속도판정, 레인 전환 속도리셋, 궁병 사거리 float, 프리셋 정비
- 기병 돌격: hasCharged → 속도 기반 판정 (풀스피드 90%+ 시 발동)
- lastSpeedBeforeAttack으로 공격 직전 속도 저장
- 레인→자유 전환 시 currentSpeed=0 리셋 (옆 행 무임 돌격 방지)
- 궁병 attackRange int→float (1.5칸 지원)
- 프리셋 3~6번 궁병 ㄴ자 배치, 3번 창병5부대
- 랜덤 배치: 함정은 col 3~4에만
- 창병 반사 12, BattleSimulator.ResetState 추가
- 근접 레인 이동 X직진 + 인접행 공격 허용

+++ prototypes/unity-prototype/Assets/Scripts/Data/UnitData.cs
@@ -25,7 +25,7 @@ public class UnitData : ScriptableObject
     public int moveEveryTick;       // N틱마다 1칸 이동 (0이면 이동 안 함)
 
     [Header("원거리 (궁병 전용)")]
-    public int attackRange;         // 사거리 (0이면 근접)
+    public float attackRange;       // 사거리 (0이면 근접, 칸 단위)
     public int shootEveryTick;      // N틱마다 1회 사격 (0이면 사격 안 함)
 
     [Header("특수 (함정 전용)")]
